+++ tests/dummy/app/controllers/ember-flexberry-dummy-suggestion-edit.js
@@ -1,6 +1,6 @@
 import Ember from 'ember';
 import BaseEditFormController from 'ember-flexberry/controllers/edit-form';
-import EditFormControllerOperationsIndicationMixin from '../mixins/edit-form-controller-operations-indication';
+import EditFormControllerOperationsIndicationMixin from 'ember-flexberry/mixins/edit-form-controller-operations-indication';
 
 export default BaseEditFormController.extend(EditFormControllerOperationsIndicationMixin, {
   /**